server/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
pymongo==4.5.0
dnspython==2.2.1
channels==4.3.0
channels-redis==4.4.0
redis==5.0.1
python-socketio==5.9.0
eventlet==0.33.3
python-dotenv==1.0.0